Добавил нлвый файл в bot/utils box.py и в нём написал функцию box, которая создаёт embed тем самым убирает повторяющийся код. Переработал команду /usercard изнутри, отличия колосальные.

+++ database/src/db/models.py
@@ -22,7 +22,7 @@ class Member(Base):
      
      member_id:  Mapped[int] = mapped_column(nullable=False, unique=True, primary_key=True)
      member_name: Mapped[str] = mapped_column(nullable=False)
-     member_cash: Mapped[int] = mapped_column(nullable=False)
+     member_cash: Mapped[float] = mapped_column(nullable=False)
      created_at:  Mapped[str] = mapped_column(nullable=False)
      joinet_at: Mapped[str] = mapped_column(nullable=False)
      avatar: Mapped[str] = mapped_column(nullable=False)
